# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Apples -- a platform where fans pitch product ideas to brands, vote them up, and preorder to trigger production.
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type IdeaStatus = typeof IdeaStatus[keyof typeof IdeaStatus];


export const IdeaStatus = {
  voting: 'voting',
  preorder: 'preorder',
  producing: 'producing',
} as const;

export interface Company {
  id: number;
  name: string;
  slug: string;
  description: string;
  /** @nullable */
  logoUrl: string | null;
  createdAt: string;
}

export interface CompanyInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  description: string;
  logoUrl?: string;
}

export interface Idea {
  id: number;
  companyId: number;
  title: string;
  description: string;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  pdfUrl: string | null;
  voteCount: number;
  preorderCount: number;
  voteThreshold: number;
  preorderThreshold: number;
  status: IdeaStatus;
  createdAt: string;
}

export interface IdeaInput {
  companyId: number;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  imageUrl?: string;
  pdfUrl?: string;
  /** @minimum 1 */
  voteThreshold?: number;
  /** @minimum 1 */
  preorderThreshold?: number;
}

export interface PreorderInput {
  /** @minLength 3 */
  email: string;
}

export interface IdeasSummary {
  totalIdeas: number;
  totalVotes: number;
  totalPreorders: number;
  ideasInProduction: number;
}

export type ListIdeasParams = {
companyId?: number;
status?: IdeaStatus;
};