# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Apples -- a platform where fans pitch product ideas to brands, vote them up, and preorder to trigger production.
 * OpenAPI spec version: 0.1.0
 */

export interface Company {
  id: number;
  name: string;
  slug: string;
  description: string;
  /** @nullable */
  logoUrl: string | null;
  createdAt: Date;
}