# lib/api-zod/src/generated/types/idea.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Apples -- a platform where fans pitch product ideas to brands, vote them up, and preorder to trigger production.
 * OpenAPI spec version: 0.1.0
 */
import type { IdeaStatus } from './ideaStatus';

export interface Idea {
  id: number;
  companyId: number;
  title: string;
  description: string;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  pdfUrl: string | null;
  voteCount: number;
  preorderCount: number;
  voteThreshold: number;
  preorderThreshold: number;
  status: IdeaStatus;
  createdAt: Date;
}