# lib/api-zod/src/generated/types/ideaInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Apples -- a platform where fans pitch product ideas to brands, vote them up, and preorder to trigger production.
 * OpenAPI spec version: 0.1.0
 */

export interface IdeaInput {
  companyId: number;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  imageUrl?: string;
  pdfUrl?: string;
  /** @minimum 1 */
  voteThreshold?: number;
  /** @minimum 1 */
  preorderThreshold?: number;
}