# lib/api-zod/src/generated/types/ideaStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Apples -- a platform where fans pitch product ideas to brands, vote them up, and preorder to trigger production.
 * OpenAPI spec version: 0.1.0
 */

export type IdeaStatus = typeof IdeaStatus[keyof typeof IdeaStatus];


export const IdeaStatus = {
  voting: 'voting',
  preorder: 'preorder',
  producing: 'producing',
} as const;