# lib/api-zod/src/generated/types/ideasSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Apples -- a platform where fans pitch product ideas to brands, vote them up, and preorder to trigger production.
 * OpenAPI spec version: 0.1.0
 */

export interface IdeasSummary {
  totalIdeas: number;
  totalVotes: number;
  totalPreorders: number;
  ideasInProduction: number;
}